+++ brayns/common/scene/Scene.cpp
@@ -87,6 +87,7 @@ void Scene::setMaterials(
         case MT_DEFAULT:
             switch( i )
             {
+                case 0: // Default
                 case 1: // Soma
                     material->setColor( Vector3f( 0.9f, 0.9f, 0.9f ));
                     break;
@@ -178,9 +179,9 @@ void Scene::buildDefault( )
 
     const uint16_t indices[6][6] =
     {
-        { 0, 1, 3, 3, 2, 0 }, // Front
+        { 5, 4, 6, 6, 7, 5 }, // Front
         { 1, 5, 7, 7, 3, 1 }, // Right
-        { 5, 4, 6, 6, 7, 5 }, // Back
+        { 0, 1, 3, 3, 2, 0 }, // Back
         { 4, 0, 2, 2, 6, 4 }, // Left
         { 0, 1, 5, 5, 4, 0 }, // Bottom
         { 2, 3, 7, 7, 6, 2 }  // Top